[#175844896] Simplified API request manager
