+++ io-pay-portal-fe/src/helper.ts
@@ -1,5 +1,13 @@
+import { fromNullable } from "fp-ts/lib/Option";
+import {
+  fromLeft,
+  taskEither,
+  TaskEither,
+  tryCatch,
+} from "fp-ts/lib/TaskEither";
 import { default as $ } from "jquery";
 import { PaymentRequestsGetResponse } from "../generated/PaymentRequestsGetResponse";
+import { apiClient } from "./api/client";
 
 export const PayDetail: ReadonlyArray<string> = [
   "importoSingoloVersamento",
@@ -9,6 +17,27 @@ export const PayDetail: ReadonlyArray<string> = [
   "enteBeneficiario",
   "spezzoniCausaleVersamento",
 ];
+
+export const getPaymentInfoTask = (
+  rptId: string
+): TaskEither<string, PaymentRequestsGetResponse> =>
+  tryCatch(
+    () =>
+      apiClient.getPaymentInfo({
+        rptId,
+      }),
+    () => "Errore recupero pagamento"
+  ).foldTaskEither(
+    (err) => fromLeft(err),
+    (errorOrResponse) =>
+      errorOrResponse.fold(
+        () => fromLeft("Errore recupero pagamento"),
+        (responseType) =>
+          responseType.status !== 200
+            ? fromLeft(`Errore recupero pagamento : ${responseType.status}`)
+            : taskEither.of(responseType.value)
+      )
+  );
 
 export const showPaymentInfo = (
   rtdId: string,
@@ -30,6 +59,8 @@ export const showPaymentInfo = (
     );
 };
 
-export const showPaymentInfoError = () => {
-  $("#error").text("Errore Validazione Pagamento");
+export const showPaymentInfoError = (errorMessage: string) => {
+  $("#error").text(
+    fromNullable(errorMessage).getOrElse("Errore Validazione Pagamento")
+  );
 };
